+++ src/modules/explorer/pages/Proposals/index.tsx
@@ -1,25 +1,23 @@
+import { Button, Grid, styled, Tooltip, Typography } from "@material-ui/core"
 import React, { useCallback, useState } from "react"
-import { Grid, styled, Typography, Button, Tooltip } from "@material-ui/core"
 
 import { useFlush } from "services/contracts/baseDAO/hooks/useFlush"
 import { useDAO } from "services/indexer/dao/hooks/useDAO"
 import { useProposals } from "services/indexer/dao/hooks/useProposals"
 import { useDAOID } from "../DAO/router"
 
-import { UserBalancesBox } from "../../components/UserBalances"
-import { ContentContainer } from "../../components/ContentContainer"
-import { ProposalsList } from "../../components/ProposalsList"
-import { DAOStatsRow } from "../../components/DAOStatsRow"
-import { Proposal, ProposalStatus } from "services/indexer/dao/mappers/proposal/types"
-import { InfoIcon } from "../../components/styled/InfoIcon"
-import { useIsProposalButtonDisabled } from "../../../../services/contracts/baseDAO/hooks/useCycleInfo"
-import { ProposalSelectionMenu } from "../../components/ProposalSelectionMenu"
-import { useDropAllExpired } from "../../../../services/contracts/baseDAO/hooks/useDropAllExpired"
-import { SmallButton } from "../../../common/SmallButton"
-import { MainButton } from "../../../common/MainButton"
 import { ProposalSelectionMenuLambda } from "modules/explorer/components/ProposalSelectionMenuLambda"
-import { useUnstakeFromAllProposals } from "services/contracts/baseDAO/hooks/useUnstakeFromAllProposals"
 import { useTezos } from "services/beacon/hooks/useTezos"
+import { useUnstakeFromAllProposals } from "services/contracts/baseDAO/hooks/useUnstakeFromAllProposals"
+import { ProposalStatus } from "services/indexer/dao/mappers/proposal/types"
+import { useIsProposalButtonDisabled } from "../../../../services/contracts/baseDAO/hooks/useCycleInfo"
+import { useDropAllExpired } from "../../../../services/contracts/baseDAO/hooks/useDropAllExpired"
+import { MainButton } from "../../../common/MainButton"
+import { SmallButton } from "../../../common/SmallButton"
+import { ContentContainer } from "../../components/ContentContainer"
+import { ProposalSelectionMenu } from "../../components/ProposalSelectionMenu"
+import { ProposalsList } from "../../components/ProposalsList"
+import { InfoIcon } from "../../components/styled/InfoIcon"
 
 const HeroContainer = styled(ContentContainer)({
   padding: "38px 38px"
@@ -57,13 +55,6 @@ export const Proposals: React.FC = () => {
   const { data: droppedProposals } = useProposals(daoId, ProposalStatus.DROPPED)
   const { account } = useTezos()
 
-  const canUnstakeVotes: boolean | undefined =
-    executedProposals &&
-    droppedProposals &&
-    executedProposals
-      .concat(droppedProposals)
-      .some(proposal => proposal.voters.find(vote => vote.address === account)?.staked)
-
   const onFlush = useCallback(async () => {
     if (executableProposals && expiredProposals && executableProposals.length && data) {
       mutate({
@@ -84,27 +75,6 @@ export const Proposals: React.FC = () => {
       return
     }
   }, [data, dropAllExpired, expiredProposals])
-
-  const onUnstakeFromAllProposals = useCallback(async () => {
-    if (droppedProposals && executedProposals && data) {
-      const allProposals = droppedProposals.concat(executedProposals)
-
-      const proposalsWithStakedTokens: Proposal[] = []
-
-      allProposals.forEach((proposal: Proposal) => {
-        const userVote = proposal.voters.find(voter => voter.address === account)
-        if (userVote && userVote.staked) {
-          proposalsWithStakedTokens.push(proposal)
-        }
-      })
-
-      unstakeFromAllProposals({
-        dao: data,
-        allProposals: proposalsWithStakedTokens.map(p => p.id)
-      })
-      return
-    }
-  }, [data, account, unstakeFromAllProposals, droppedProposals, executedProposals])
 
   const toggleProposalModal = () => {
     switch (data?.data.dao_type?.name) {
@@ -147,19 +117,6 @@ export const Proposals: React.FC = () => {
                     disabled={!expiredProposals || !expiredProposals.length}
                   >
                     Drop All Expired
-                  </DropButton>
-                  <Tooltip placement="bottom" title="Drop all expired proposals">
-                    <InfoIcon color="secondary" />
-                  </Tooltip>
-                </Grid>
-                <Grid item>
-                  <DropButton
-                    variant="contained"
-                    color="secondary"
-                    onClick={onUnstakeFromAllProposals}
-                    disabled={!canUnstakeVotes}
-                  >
-                    Unstake From Proposals
                   </DropButton>
                   <Tooltip placement="bottom" title="Drop all expired proposals">
                     <InfoIcon color="secondary" />
